✨ feat: 添加启动页；移除过时页面
