--- miniprogram/pages/issue/index.json
@@ -1,4 +1,0 @@
-{
-  "enablePullDownRefresh": true,
-  "usingComponents": {}
-}
