--- miniprogram/pages/login/index.wxss
@@ -1,4 +1,0 @@
-/* pages/login/index.wxss */
-page {
-  
-}
