--- miniprogram/pages/profile/index.json
@@ -1,4 +1,0 @@
-{
-  "navigationBarTitleText": "",
-  "usingComponents": {}
-}
