--- miniprogram/pages/profile/index.wxml
@@ -1,15 +1,0 @@
-<!-- pages/profile/index.wxml -->
-
-<ui-sys title="个人资料" data-weui-theme="{{theme}}">
-  <van-notice-bar left-icon="volume-o" text="头像昵称填写功能，正在优化，马上写完，信我。" />
-  <button class="avatar-wrapper" open-type="chooseAvatar" bind:chooseavatar="onChooseAvatar">
-    <image class="avatar" src="{{avatarUrl}}"></image>
-  </button> 
-  <mp-form>
-    <mp-cells>
-      <mp-cell title="昵称">
-        <input type="nickname" class="weui-input" placeholder="请输入昵称"/>
-      </mp-cell>
-    </mp-cells>
-  </mp-form>
-</ui-sys>
