--- miniprogram/pages/profile/index.wxss
@@ -1,18 +1,0 @@
-/* pages/profile/index.wxss */
-.avatar-wrapper {
-  padding: 0;
-  width: 56px !important;
-  border-radius: 8px;
-  margin-top: 40px;
-  margin-bottom: 40px;
-}
-
-.avatar {
-  display: block;
-  width: 56px;
-  height: 56px;
-}
-
-.container {
-  display: flex;
-}
