+++ miniprogram/pages/start/index.scss
@@ -1,0 +1,1 @@
+/* pages/start/index.wxss */
