+++ miniprogram/pages/start/index.wxml
@@ -1,0 +1,8 @@
+<!--pages/start/index.wxml-->
+<ui-sys img="/static/img/bg2.svg" loading="{{false}}">
+	<view class="ui-container flex-column flex-center" style="height: 90vh;">
+		<ui-avatar ui="sl ellipse" src="/static/logo.png" />
+		<!-- 网络错误 -->
+		<van-empty class="mt-6" wx:if="{{netError}}" image="network" description="暂停服务" />
+	</view>
+</ui-sys>
